[utils] one_conn_awk_bench: add progress bar
Useful to track when server stop reponding because or segfaulted main
server of multi-thread server, that was introduced first in 0a01f795595.

bc-benchmark not cause segfault server, but current one does
(one_conn_awk_bench)
Seems that because of one connection to server (in current script), and
it try to reuse one session for many threads, which is not good.