[kernel] reimplement multi-thread server using pool of io services.
Previous implementation has a bug, with one connection to server, it try
to reuse one session for many threads, which is not good. (see 27faeab692)

So I decide to use pool of io_service based on
boost_asio/example/http/server2/io_service_pool.hpp
And it fix the initial issue.

First multi-thread server was introduced in 0a01f7955.

Speed of this implemention like in previous (multi-thread one) see
8d27bb52a8c for more info.