[test] travis-ci: try to build on g++ 4.7 instead of clang 3.2 (HACK)
We can't use clang because libstdc++4.7 contains a broken definition of
std::common_type, and the hack for it is only in clang 3.2~rc1-1~exp1.
See also:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=666539
http://llvm.org/bugs/show_bug.cgi?id=12893

So try to build using g++ 4.7, that support c++11.
When travis-ci/travis-core#195 issue will be resolved we need to
"revert" this commit.

See also c696a29429b7