server: use per-thread base for session objects (to aovid races)
Before this patch boostcached with particular options exited from main
event loop, even if we don't asked to.

Steps to reproduce:
  $ boostcached -w1 -vvv
  $ bc-benchmark -s .cmake/boostcached.sock

After benchmark finished server existed, without preparing to stop,
and appropriate message will be written (about routine is not
prepared for this).

I believe that this is because we use main base (created in the main
thread), in our routines, which are separate threads by it's nature.
And this is very bad, since it can harm event_base internals.

After this patch all seems to be okay, except that no response for first
request from server with -vvv, but with -vvvv we have a response. (Ugh.)