utils/sanitizers: drop CMakeCache from the same as CMakeLists to avoid errors
If somebody build project without extra subdir for cmake generated stuff,
then current script will pick up cache from the same directory where
CMakeLists.txt is, and will fail because of cached *COMPILER* variables.
So just drop it and that's it.

Error messages like this:
  You have changed variables that require your cache to be deleted.
  Configure will be re-run and you may have to reset some variables.
  The following variables have changed:
  CMAKE_CXX_COMPILER= clang++
  CMAKE_C_COMPILER= clang

And also because of this, detection of the compiler family will be wrong and
wrong flags will be checked (gcc and clang have different needs for thread
sanitizer).

CI: https://travis-ci.org/azat/boostcache/builds/24924628