build: use OVERWRITE for AddCompilerFlags macro (updated)
New AddCompilerFlags() is more generic, which will allow not to check
flags for not you build type, but if you want to you must use FORCE, but
this patch don't add this, so it change the behaviour which compiler
flags will be validated.
And what about OVERWRITE, this is used if you need to pass flags for
linker stage not the compiler one.